api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
requests==2.32.3
python-dateutil==2.9.0
websockets==12.0
python-multipart==0.0.9
